[Core][CoreBundle] Move shipping category from Product to productVariant
