+++ src/Sylius/Bundle/AdminBundle/Resources/views/Product/Tab/_details.html.twig
@@ -19,6 +19,7 @@
                         {% endif %}
                     </div>
                     {{ form_row(form.variant.price) }}
+                    {{ form_row(form.variant.shippingCategory) }}
                     {{ form_row(form.variant.onHand) }}
                     {{ form_row(form.variant.tracked) }}
                 {% else %}
