+++ src/Sylius/Bundle/AdminBundle/Resources/views/ProductVariant/Tab/_details.html.twig
@@ -18,6 +18,9 @@
                     {% endif %}
                 </div>
             </div>
+            <div class="two fields">
+                {{ form_row(form.shippingCategory) }}
+            </div>
             {% for name, calculatorConfigurationPrototype in form.vars.prototypes %}
                 <div id="{{ form.pricingCalculator.vars.id }}_{{ name }}" data-container=".configuration"
                      data-prototype="{{ form_widget(calculatorConfigurationPrototype)|e }}">
